Add `ElasticIPAddress` CRD (#39)
Issue #, if available: https://github.com/aws-controllers-k8s/community/issues/1218

Description of changes:
Creates the `ElasticIPAddress` custom resource, from `AllocateAddress` and `DeallocateAddress`. This API uses either `AllocationID` or `PublicIP` as the identifier depending on whether the domain is set to `"vpc"` (for EC2-VPC) or otherwise (for EC2-Classic), respectively. For the adoption, I have chosen to set `AllocationID` as the primary key for now, since EC2-VPC is the recommended method for managing resources in EC2.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.