Add route table assocations to `Subnet` (#45)
Closes https://github.com/aws-controllers-k8s/community/issues/1219

Description of changes:
Generates the `Subnet` CRD with a new custom field `RouteTables`, which represent an array of route table IDs. Route tables are associated with the `Subnet` using the `AssociateRouteTable` API call, and then updated by deleting the association (using `DisassociateRouteTable`) and re-associating. 

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.