VPC attachment for `InternetGateway` (#48)
Closes https://github.com/aws-controllers-k8s/community/issues/1216

Description of changes:
Add support for attaching an `InternetGateway` to a `VPC`. The EC2 API exposes a list of `Attachments` when you describe `InternetGateway`, but manual testing shows that you can only attach a gateway to a single VPC - so this is treated as a list of at most one element. I had to create a new VPC for integration tests, since the default VPC we use for all of the other tests already contains an IGW, and you can have at most 1 per VPC.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.