Fix Internet Gateway terminal codes (#68)
Issue #, if available: [1362](https://github.com/aws-controllers-k8s/community/issues/1362)

Description of changes: 

- Removed `InvalidVpcId.Malformed` error from terminal codes for **Internet Gateway** resource.
- Removed Integration tests for the `InvalidVpcId.Malformed` terminal code.

Initially, I thought `InvalidVpcId.Malformed` should be included in the terminal codes. Later, observed that Internet Gateway resource being created after adding the integration tests for the malformed vpc id terminal code. This is expected as per our code because initially the resource will be created with no vpc attachment and later if we pass VPC ID, it will be attached to the resource. After discussions and local testing I observed that if you add malformed VPC ID it will create a resource and if you try to update the VPC ID to a valid one, it will create a new resource and leave the initial one in dangling state that cannot be deleted, so we decided not to include malformed VPC ID  for Internet Gateway resource.


By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.