Update ACK runtime, code-generator `v0.20.1` & aws-sdk-go to `v1.44.93` (#93)
Issue #, if available: https://github.com/aws-controllers-k8s/community/issues/1203

Description of changes: Updates *runtime* and *aws-sdk-go* dependencies to use the latest aws models/shapes and ignores newly identified resources. The latest models are needed to align controller functionality with that of the console.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.