Fix: EC2 controller for vpcendpoint does not update status
Issue : aws-controllers-k8s/community#2075

Description of changes:
The issue is that vpce CR status is not updated. This happens because when vpce is created, it takes around a minute for it to be created in aws. Due to this, it's status is kept pending. But there is no sync enforced on the reconciler. Due to this, reconciliation happens after 10h default time. At this time, status would be set correctly.

This fix enforces reconciliation to make sure CR is updated correctly when vpce creation in aws is successful.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.