Enhance `RouteTable` CRD sync routes synchronisation (#199)
- Update `syncRoutes` method to handle errors from `excludeAWSRoute`
- Reorder route deletion and addition in `syncRoutes` for proper sync
- Modify `excludeAWSRoute` to return an error and rename variables for clarity
- Move tags sync before routes sync in `customUpdateRouteTable`
- Update `sdkCreate` to use a copy of the RouteTable for creating routes
- Minor formatting and comment updates in template files

The changes in this commit enhance the synchronization of routes for the
RouteTable Custom Resource Definition (CRD). The `syncRoutes` method now handles
errors returned from the `excludeAWSRoute` function and reorders the deletion
and addition of routes for proper synchronization. The `excludeAWSRoute`
function is modified to return an error and has some variable renaming for
better clarity...

In the `customUpdateRouteTable` method the tags synchronization is moved before
the routes synchronization to ensure tagss are updated before route changes.

The `sdkCreate` method now uses a deep copy of the RouteTable when creating
routes to avoid modifying the original desired state

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.