fix: only force syncing a VPC endpoint when in pending state (#198)
Context: https://github.com/aws-controllers-k8s/ec2-controller/pull/191#discussion_r1627514799

A synced VPC endpoint should only be forced to re-sync when it's in `pending` state. For other states like `pendingAcceptance`, `rejected`, `failed`, `expired`, it could be the final state of the endpoint and the sync should be considered completed.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.