fix(network_acl): Preserve desired entries order after update (#208)
This patch addresses an issue where the EC2 API may return entries
in a different order than specified. Causing the controller to go
change the desired state spec.

Now, after updating the NACL, we fetch the latest resource status
but we preserve the order of entries.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.