Fix: Add option to delete default SG rule created upon VPC creation (#212)
Issue #, if available:

Description of changes:
When a VPC is created, it also creates 'default' security group. The 'default' security group has widely open egress rules. We need to have an option to delete this autocreated rules from 'default' security group. The 'default' security group itself can not be deleted.

Discussion:
https://kubernetes.slack.com/archives/C0402D8JJS1/p1720560499642019

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.