Fix VPC DNS attribute handling
When creating a VPC there were errors in the K8s logs of the controller.
From CloudTrail perspective the following could be seen:

    ..
    "eventSource": "ec2.amazonaws.com",
    "eventName": "ModifyVpcAttribute",
    ...
    "errorCode": "Client.InvalidParameterCombination",
    "errorMessage": "No attributes specified.",
    ...
    "requestParameters": {
        "vpcId": "vpc-xxxxxxxxxxxxxxxxx"
    },
    ..

This is giving the `vpcID` as parameter, but no attribute to change.

Investigating I found the following. The controller wants to modify the
attributes, as the delta contains the following drift (for debugging I
printed the difference in the delta):

    {"msg":"difference","path":{"Parts":["Spec","EnableDNSSupport"]},"a":null,"b":true}
    {"msg":"difference","path":{"Parts":["Spec","EnableDNSHostnames"]},"a":null,"b":false}

So the desired state ("a") doesn't specify these attributes, but the
latest state ("b") has them (the values are the defaults as per [1]).
Therefore it wants to modify the attributes. As there is no
configuration in desired, the attribute is not added to the inputs (see
[2] and [3]), resulting in the above error.

Therefore adding the nil-check seems to be the best way to handle this.

Fixes #1826.

[1] https://docs.aws.amazon.com/vpc/latest/userguide/AmazonDNS-concepts.html#vpc-dns-support
[2] https://github.com/aws-controllers-k8s/ec2-controller/blob/b18aa397313cef18db06d2872cdcedd8254aa283/pkg/resource/vpc/hooks.go#L76
[3] https://github.com/aws-controllers-k8s/ec2-controller/blob/b18aa397313cef18db06d2872cdcedd8254aa283/pkg/resource/vpc/hooks.go#L91