Improve the DNS attribute default handling (#226)
This improves #225.

Using a custom pre-compare function the defaults can be set accordingly. This is now added with the defaults set to what the VPC defaults are. The `disallowSecurityGroupDefaultRules` option is also handled here, as it also has the issue describe in aws-controllers-k8s/community#1826.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.