Make routes order independent
In the past the routes were not order independent. Whenever the API
would return a different order than specified, the routes would be
deleted and recreated. The `getMatchingRoute` code also returned the
first route that would have a matching field like the
VPCPeeringConnectionId. This resulted in routes not being stable.

Now the `customPreCompare` handles routes properly, i.e. is order
independent. Also the `syncRoutes` uses the computed diff, to create or
delete the according routes.